Move parsing of Luna client header files to gem installation
In the pkcs11 base gem, we ship the PKCS#11 standard header files and link to the PKCS#11 ABI (not the API).

Due to copyright constraints we're not allowed to distribute Luna client header files in the gem as we do in the pkcs11 gem.
Instead we compile with the header files provided by the Luna client package, so we're using the API (not the ABI).

So far we distibuted the struct names and constant name/value pairs extracted from the header files at build time.
This however leads to inconsistencies between generated files at build time and the header files at install time, if the Luna client versions differ.
Moving the parsers to extconf.rb ensures that generated structs and constants fit to the header files that are included by the compiler.